netlog/requirements.txt
aiohttp==3.7.2
ansicon==1.89.0
arrow==0.17.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
blessed==1.17.11
certifi==2020.11.8
chardet==3.0.4
Django==3.1.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.2
django-datetime-widget==0.9.3
django-extensions==3.1.0
django-geoip2-extras==1.2.1
django-picklefield==3.0.1
django-redis==4.12.1
django-widget-tweaks==1.4.8
geoip2==3.0.0
graphviz==0.15
idna==2.10
jinxed==1.0.1
maxminddb==2.0.3
multidict==5.0.2
psycopg2==2.8.6
pydotplus==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-env-utils==0.4.1
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2020.1
redis==3.5.3
requests==2.25.0
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
yarl==1.6.3
